test: split whatsapp setup surface coverage
